Allow propagating action output in context for chaining actions
With this commit, if the action output is a map it gets merged
to the context; to prevent that, the user can add the `^:ignore`
metadata entry to map returned by the action handler.

The intent of this is to enable implementing actions that rely
on the computation of previous action steps;
before this patch such behaviour was only possible by
defining A as a "gatherer", which effectively broke the
implicit semantics of the library: 1) no result on
"outputs" for such action gatherer and 2) using a
gatherer for a procedure that may introduce state changes.

This is a behaviour change and may impact any existent code that
has actions in sequence; the context value that gets propagated
to subsequent entries may end up having entries overriden by
previous actions.